make aerodynamic surfaces into components and surface effectors into subcomponents
