+++ include/simulation/aerodynamics/public/manager.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "simulation/aerodynamics/public/data/types.hpp"
+#include "simulation/aerodynamics/public/components/surface.hpp"
 
 #include <vector>
 
@@ -10,7 +10,6 @@ namespace aerodynamics
 		std::vector<Surface> surfaces;
 
 		AerodynamicsManager(std::vector<Surface> s);
-		void compute_surface_geometry();
 		template <typename T> AerodynamicsManagerOutput_T<T> step(const AerodynamicsManagerInput_T<T>& input);
 	};
