+++ include/simulation/aerodynamics/public/manager.tpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "simulation/aerodynamics/public/detail/loads.hpp"
+#include "simulation/aerodynamics/public/manager.hpp"
 
 namespace aerodynamics
 {
@@ -8,10 +8,18 @@ namespace aerodynamics
 	AerodynamicsManagerOutput_T<T> AerodynamicsManager::step(
 		const AerodynamicsManagerInput_T<T>& input)
 	{
-		return {
-			.WB_aerodynamic =
-				step_aero_forces_moments_T<T>(surfaces, input.pB_GB, input.twist, input.atm, input.u, input.windB)
+		const SurfaceInput_T<T> surface_input{
+			.pB_GB = input.pB_GB, .twist = input.twist, .atm = input.atm, .u = input.u, .windB = input.windB
 		};
+
+		dynamics::Wrench_T<T> WB_aerodynamic;
+		for (Surface& surface : surfaces) {
+			const dynamics::Wrench_T<T> WB_surface = surface.step<T>(surface_input);
+			WB_aerodynamic.F += WB_surface.F;
+			WB_aerodynamic.M += WB_surface.M;
+		}
+
+		return {.WB_aerodynamic = WB_aerodynamic};
 	}
 
 } // namespace aerodynamics
