+++ src/simulation/aerodynamics/public/components/surface.cpp
@@ -1,0 +1,12 @@
+#include "simulation/aerodynamics/public/components/surface.hpp"
+
+namespace aerodynamics
+{
+
+	void Surface::compute_geometry()
+	{
+		area = chord * span;
+		AR = span / chord;
+	}
+
+} // namespace aerodynamics
