+++ src/simulation/aerodynamics/public/manager.cpp
@@ -9,14 +9,8 @@ namespace aerodynamics
 		std::vector<Surface> s)
 		: surfaces(std::move(s))
 	{
-		compute_surface_geometry();
-	}
-
-	void AerodynamicsManager::compute_surface_geometry()
-	{
-		for (Surface& s : surfaces) {
-			s.area = s.chord * s.span;
-			s.AR = s.span / s.chord;
+		for (Surface& surface : surfaces) {
+			surface.compute_geometry();
 		}
 	}
